refactor(LottoTicket) LottoTicket과 Numbers의 상속관계 제거
불필요한 상속관계를 제거하고 Numbers 클래스를 삭제
