+++ src/main/java/lotto/domain/LottoTicket.java
@@ -7,23 +7,38 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-public class LottoTicket extends Numbers {
+public class LottoTicket {
+
+    private final Set<LottoNumber> lottoNumbers;
 
     public LottoTicket(Set<LottoNumber> lottoNumbers) {
-        super(lottoNumbers);
-        if (lottoNumbers.size() != 6) {
-            throw new RuntimeException("로또 번호는 중복되지 않는 6개의 수로 구성되어야 합니다.");
-        }
+        validateLottoTicket(lottoNumbers);
+        this.lottoNumbers = lottoNumbers;
     }
 
     public LottoTicket(List<LottoNumber> lottoNumbers) {
         this(new HashSet<>(lottoNumbers));
     }
 
-    public TicketDto toDto() {
-        return new TicketDto(lottoNumbers.stream()
-            .map(LottoNumber::toInteger)
-            .collect(Collectors.toList()));
+    public int compare (LottoTicket otherNumbers) {
+        return (int) lottoNumbers.stream()
+            .filter(otherNumbers::contains)
+            .count();
     }
 
+    public boolean contains(LottoNumber lottoNumber) {
+        return lottoNumbers.contains(lottoNumber);
+    }
+
+    public TicketDto toDto() {
+        return new TicketDto(lottoNumbers.stream()
+                .map(LottoNumber::toInteger)
+                .collect(Collectors.toList()));
+    }
+
+    private static void validateLottoTicket(Set<LottoNumber> lottoNumbers) {
+        if (lottoNumbers.size() != 6) {
+            throw new RuntimeException("로또 번호는 중복되지 않는 6개의 수로 구성되어야 합니다.");
+        }
+    }
 }
